+++ src/parsers/parseAnnotations.js
@@ -21,6 +21,9 @@ export default function parseAnnotations(version: string, content: string): CLDR
 
   xml('annotation').each((i, rawRow) => {
     const row = xml(rawRow);
+
+    // Variation selectors are not present in the locale files
+    // So lets just strip unnecessary codepoints
     const hexcode = fromUnicodeToHexcode(row.attr('cp'));
 
     if (!data[hexcode]) {